service/src/main/java/com/example/service/impl/UserServiceImpl.java
package com.example.service.impl;

import com.example.database.entity.User;
import com.example.database.repository.UserRepository;
import com.example.service.dto.UserDto;
import com.example.service.interfaces.UserService;
import jakarta.transaction.Transactional;
import lombok.RequiredArgsConstructor;
import org.modelmapper.ModelMapper;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageImpl;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;
import java.util.stream.Collectors;

@Service
@Transactional
@RequiredArgsConstructor
public class UserServiceImpl implements UserService {
    private final UserRepository userRepository;
    private final ModelMapper modelMapper;


    @Override
    public Optional<UserDto> create(UserDto userDto) {
        User user = modelMapper.map(userDto, User.class);
        user = userRepository.save(user);
        return Optional.of(user).map(user1 -> modelMapper.map(user1, UserDto.class));
    }

    @Override
    public Page<UserDto> findAll(Pageable pageable) {
        Page<User> userPage = userRepository.findAll(pageable);
        List<UserDto> userDtoList = userPage
                .stream()
                .map(user -> modelMapper.map(user, UserDto.class))
                .collect(Collectors.toList());
        return new PageImpl<>(userDtoList, pageable, userPage.getTotalElements());
    }


}
